Finds the glomerulus associated with a PN

Given a vector of skeleton IDs, this will find the glomerului associated with PNs, based on annotations in CATMAID.

find_glomeruli(skids, exclude = NULL, method = c("annotation", "name"),
  ...)

Arguments

skids
Required; an integer or character vector of skeleton IDs

exclude
Optional; a vector of glomeruli to exclude (in the format X for glomerulus X, or the full annotation for unknown glomerulus N)

method
Optional; indicating whether to use the neuron 'name' or glomerulus 'annotation's.

...
Additional arguments passed to catmaid_* functions. Use this to specify conn, pid or similar low level arguments for the catmaid_query.

Value

A character vector of glomeruli names

Details

Requires the PN to be annotated with glomerulus X or unknown glomerulus N in CATMAID. Annotations of the form glomerulus X will be prioritised over unknown glomerulus N, and if there are multiples they will be joined together with a forward slash. If there are no results with either of these annotations, the string "unknown" will be returned in place of a glomerulus. Note that annotations of the form glomerulus X right|left are not considered.

Examples

find_glomeruli('WTPN2017_uPN_right')

#>   [1] "VA6"      "VL2a"     "VC1"      "VC3m"     "DL2v"     "DM5"     
#>   [7] "VC3m"     "VA5"      "VL2p"     "VC3l/VCy" "VA2"      "VM5v"    
#>  [13] "VA4"      "VM4"      "DL4"      "VM7v"     "D"        "VM1"     
#>  [19] "DP1l"     "VM3"      "DA1"      "DM5"      "V"        "DM3"     
#>  [25] "VM5v"     "DL2d"     "DM1"      "VA7m"     "DC3"      "DL3"     
#>  [31] "D"        "VM5d"     "VA7m"     "VA1d"     "DA4m"     "VC4"     
#>  [37] "DA2"      "VM5d"     "VC5"      "DL5"      "DM4"      "VM7d"    
#>  [43] "D"        "DA4l"     "VM3"      "VC3l"     "VA1d"     "VC4"     
#>  [49] "VM7v"     "VC2"      "DC4"      "VC3l"     "VM5d"     "VA1v"    
#>  [55] "VM2"      "VA1v"     "DA2"      "VM5v"     "VM2"      "VC4"     
#>  [61] "VA1v"     "DM2"      "DL2v"     "VCx/VM4"  "VC5"      "VM4"     
#>  [67] "VC5"      "DC3"      "VA1v"     "DM6"      "DA1"      "VA3"     
#>  [73] "DA1"      "DL2d"     "DL2d"     "DL2d"     "DA3"      "DA1"     
#>  [79] "VA5"      "VA5"      "DA1"      "DM5"      "DC2"      "VA7l"    
#>  [85] "DC3"      "DA2"      "DA2"      "DL3"      "VL1"      "DP1m"    
#>  [91] "DM6"      "DA1"      "DL2v"     "VM5d"     "DA3"      "DC2"     
#>  [97] "DM2"      "DL2v"     "DM6"      "VL1"      "DL3"      "VA3"     
#> [103] "DL3"      "VA7m"     "V"        "VC3m"     "VA1d"     "DL3"     
#> [109] "DL2d"     "DA1"      "DC1"      "VM1"      "DP1m"     "VL2p"    
#> [115] "VA1v"     "DL1"      "DA2"      "DA1"      "VL2a"     "DC3"     
#> [121] "VA1v"     "VL1"      "DA1"      "VL2a"     "DP1m"    

find_glomeruli('WTPN2017_uPN_right', exclude=c('VCx','VCy', 'VP1l'))

#>   [1] "VA6"  "VL2a" "VC1"  "VC3m" "DL2v" "DM5"  "VC3m" "VA5"  "VL2p" "VC3l"
#>  [11] "VA2"  "VM5v" "VA4"  "VM4"  "DL4"  "VM7v" "D"    "VM1"  "DP1l" "VM3" 
#>  [21] "DA1"  "DM5"  "V"    "DM3"  "VM5v" "DL2d" "DM1"  "VA7m" "DC3"  "DL3" 
#>  [31] "D"    "VM5d" "VA7m" "VA1d" "DA4m" "VC4"  "DA2"  "VM5d" "VC5"  "DL5" 
#>  [41] "DM4"  "VM7d" "D"    "DA4l" "VM3"  "VC3l" "VA1d" "VC4"  "VM7v" "VC2" 
#>  [51] "DC4"  "VC3l" "VM5d" "VA1v" "VM2"  "VA1v" "DA2"  "VM5v" "VM2"  "VC4" 
#>  [61] "VA1v" "DM2"  "DL2v" "VM4"  "VC5"  "VM4"  "VC5"  "DC3"  "VA1v" "DM6" 
#>  [71] "DA1"  "VA3"  "DA1"  "DL2d" "DL2d" "DL2d" "DA3"  "DA1"  "VA5"  "VA5" 
#>  [81] "DA1"  "DM5"  "DC2"  "VA7l" "DC3"  "DA2"  "DA2"  "DL3"  "VL1"  "DP1m"
#>  [91] "DM6"  "DA1"  "DL2v" "VM5d" "DA3"  "DC2"  "DM2"  "DL2v" "DM6"  "VL1" 
#> [101] "DL3"  "VA3"  "DL3"  "VA7m" "V"    "VC3m" "VA1d" "DL3"  "DL2d" "DA1" 
#> [111] "DC1"  "VM1"  "DP1m" "VL2p" "VA1v" "DL1"  "DA2"  "DA1"  "VL2a" "DC3" 
#> [121] "VA1v" "VL1"  "DA1"  "VL2a" "DP1m"

find_glomeruli('WTPN2017_uPN_right', method = 'name')

#>   [1] "VA6"  "VL2a" "VC1"  "VC3m" "DL2v" "DM5"  "VC3m" "VA5"  "VL2p" "VC3l"
#>  [11] "VA2"  "VM5v" "VA4"  "VM4"  "DL4"  "VM7v" "D"    "VM1"  "DP1l" "VM3" 
#>  [21] "DA1"  "DM5"  "V"    "DM3"  "VM5v" "DL2d" "DM1"  "VA7m" "DC3"  "DL3" 
#>  [31] "D"    "VM5d" "VA7m" "VA1d" "DA4m" "VC4"  "DA2"  "VM5d" "VC5"  "DL5" 
#>  [41] "DM4"  "VM7d" "D"    "DA4l" "VM3"  "VC3l" "VA1d" "VC4"  "VM7v" "VC2" 
#>  [51] "DC4"  "VC3l" "VM5d" "VA1v" "VM2"  "VA1v" "DA2"  "VM5v" "VM2"  "VC4" 
#>  [61] "VA1v" "DM2"  "DL2v" "VM4"  "VC5"  "VM4"  "VC5"  "DC3"  "VA1v" "DM6" 
#>  [71] "DA1"  "VA3"  "DA1"  "DL2d" "DL2d" "DL2d" "DA3"  "DA1"  "VA5"  "VA5" 
#>  [81] "DA1"  "DM5"  "DC2"  "VA7l" "DC3"  "DA2"  "DA2"  "DL3"  "VL1"  "DP1m"
#>  [91] "DM6"  "DA1"  "DL2v" "VM5d" "DA3"  "DC2"  "DM2"  "DL2v" "DM6"  "VL1" 
#> [101] "DL3"  "VA3"  "DL3"  "VA7m" "V"    "VC3m" "VA1d" "DL3"  "DL2d" "DA1" 
#> [111] "DC1"  "VM1"  "DP1m" "VL2p" "VA1v" "DL1"  "DA2"  "DA1"  "VL2a" "DC3" 
#> [121] "VA1v" "VL1"  "DA1"  "VL2a" "DP1m"